feat: Add members and groups modules (#3)
Modules:

1. Membership management
Admin users should be able to add members of the team. Each member can be assigned to a particular group (for example, guitarists, singers, pianists, etc.). Non PII can be added as metadata of the member, mostly for contact information.

2. Groups management
Admins can create and update groups, as well as member assignments between groups. Note that a member can belong to zero-to-many groups.

For the REST API, let's use a yaml openapi specification for all the endpoints, requests and responses. Also add migration scripts for the database needs.